:white_check_mark: Add `test_patch_classifier`
